fix: use correct broadcast address
