+++ orchestrator/client/node.py
@@ -3,6 +3,7 @@ import signal
 import socket
 import subprocess
 from orchestrator.shared import *
+from orchestrator.utils import get_broadcast_ip
 
 class Node:
     def __init__(self, socket, ip):
@@ -10,8 +11,11 @@ class Node:
         self.__socket = socket
         self.__is_master = False
         self.__pid = -1
+        self.__broadcast_ip = get_broadcast_ip()
 
     def start(self):
+        print("Received start command")
+
         if self.__is_master:
             make_cmd = "run_rcv_joystick_pi"
         else:
@@ -21,6 +25,7 @@ class Node:
         self.__pid = proc.pid
 
     def stop(self):
+        print("Received stop command")
         if self.__pid == -1:
             return
 
@@ -28,10 +33,13 @@ class Node:
         self.__pid = -1
 
     def send_heartbeat(self):
-        self.__socket.sendto(str.encode(MSG_CMD_HEARTBEAT), ("255.255.255.255", SOCKET_PORT))
+        print("Sending heartbeat")
+        self.__socket.sendto(str.encode(MSG_CMD_HEARTBEAT), (self.__broadcast_ip, SOCKET_PORT))
 
     def handle_message(self, msg):
-        (cmd, data) = msg
+        cmd = msg[0]
+        data = "" if len(msg) == 1 else msg[1]
+
         if msg == MSG_CMD_SET_MASTER:
             self.__is_master = data == self.__ip
         elif msg == MSG_CMD_START:
