+++ orchestrator/server.py
@@ -1,7 +1,9 @@
 from orchestrator.shared import *
+from orchestrator.utils import get_broadcast_ip
 
 class Server:
     def __init__(self, socket):
+        self.__broadcast_ip = get_broadcast_ip()
         self.__master_node = None
         self.__is_running = False
         self.__socket = socket
@@ -19,6 +21,7 @@ class Server:
         cmd = msg[0]
         if cmd == MSG_CMD_HEARTBEAT:
             if data not in self.__nodes:
+                print("New node {data} connected")
                 self.__nodes.append(data)
 
     def handle_user_cmd(self, msg):
@@ -31,10 +34,10 @@ class Server:
                 return
             self.__socket.sendto(str.encode(cmd), (data, SOCKET_PORT))
         elif cmd == MSG_CMD_START:
-            self.__socket.sendto(str.encode(cmd), ("255.255.255.255", SOCKET_PORT))
+            self.__socket.sendto(str.encode(cmd), (self.__broadcast_ip, SOCKET_PORT))
             self.__is_running = True
         elif cmd == MSG_CMD_STOP:
-            self.__socket.sendto(str.encode(cmd), ("255.255.255.255", SOCKET_PORT))
+            self.__socket.sendto(str.encode(cmd), (self.__broadcast_ip, SOCKET_PORT))
             self.__is_running = False
         else:
             print("Unsupported command")
